refactor: type.ts로 분리
