+++ client/src/components/SubmitDialog/types.ts
@@ -1,0 +1,5 @@
+export interface SubmitDialogProps {
+    open: boolean;
+    onConfirm: () => void;
+    onClose: () => void;
+}
